refactor(storage): split storage.go by responsibility
